requirements.txt
requests
pandas
pydantic
fastapi
uvicorn
fastmcp
openai
